+++ src/consumer/component-ops/component-writer.js
@@ -15,6 +15,8 @@ import type { PathOsBasedRelative } from '../../utils/path';
 import { preparePackageJsonToWrite } from '../component/package-json-utils';
 import DataToPersist from '../component/sources/data-to-persist';
 import RemovePath from '../component/sources/remove-path';
+import BitMap from '../bit-map/bit-map';
+import ConfigDir from '../bit-map/config-dir';
 import EnvExtension from '../../extensions/env-extension';
 import ComponentConfig from '../config/component-config';
 import { populateEnvFilesToWrite } from './eject-conf';
@@ -26,8 +28,10 @@ export type ComponentWriterProps = {
   configDir?: string,
   writePackageJson?: boolean,
   override?: boolean,
+  isolated?: boolean,
   origin: ComponentOrigin,
-  consumer: Consumer,
+  consumer: ?Consumer,
+  bitMap: BitMap,
   writeBitDependencies?: boolean,
   deleteBitDirContent?: boolean,
   existingComponentMap?: ComponentMap,
@@ -41,8 +45,10 @@ export default class ComponentWriter {
   configDir: ?string;
   writePackageJson: boolean;
   override: boolean;
+  isolated: ?boolean;
   origin: ComponentOrigin;
-  consumer: Consumer;
+  consumer: ?Consumer;
+  bitMap: BitMap;
   writeBitDependencies: boolean;
   deleteBitDirContent: ?boolean;
   componentMap: ComponentMap;
@@ -55,8 +61,10 @@ export default class ComponentWriter {
     configDir,
     writePackageJson = true,
     override = true,
+    isolated = false,
     origin,
     consumer,
+    bitMap,
     writeBitDependencies = false,
     deleteBitDirContent,
     existingComponentMap,
@@ -68,8 +76,10 @@ export default class ComponentWriter {
     this.configDir = configDir;
     this.writePackageJson = writePackageJson;
     this.override = override;
+    this.isolated = isolated;
     this.origin = origin;
     this.consumer = consumer;
+    this.bitMap = bitMap;
     this.writeBitDependencies = writeBitDependencies;
     this.deleteBitDirContent = deleteBitDirContent;
     this.existingComponentMap = existingComponentMap;
@@ -89,7 +99,9 @@ export default class ComponentWriter {
    * when a component is not new, write the files according to the paths in .bitmap.
    */
   async write(): Promise<Component> {
+    if (!this.consumer) throw new Error('ComponentWriter.write expect to have a consumer');
     await this.populateComponentsFilesToWrite();
+    // $FlowFixMe consumer is set
     this.component.dataToPersist.addBasePath(this.consumer.getPath());
     await this.component.dataToPersist.persistAllToFS();
     return this.component;
@@ -125,13 +137,13 @@ export default class ComponentWriter {
     if (dists) this.component.dataToPersist.merge(dists);
     if (this.writeConfig && this.consumer) {
       const resolvedConfigDir = this.configDir || this.consumer.dirStructure.ejectedEnvsDirStructure;
-      const configToWrite = await this.component.getConfigToWrite(this.consumer, resolvedConfigDir);
+      const configToWrite = await this.component.getConfigToWrite(this.consumer, this.bitMap, resolvedConfigDir);
       this.component.dataToPersist.merge(configToWrite.dataToPersist);
     }
     // make sure the project's package.json is not overridden by Bit
     // If a consumer is of isolated env it's ok to override the root package.json (used by the env installation
     // of compilers / testers / extensions)
-    if (this.writePackageJson && (this.consumer.isolated || this.writeToPath !== this.consumer.getPath())) {
+    if (this.writePackageJson && ((this.consumer && this.consumer.isolated) || this.writeToPath !== '.')) {
       const { packageJson, distPackageJson } = preparePackageJsonToWrite(
         this.consumer,
         this.component,
@@ -169,7 +181,7 @@ export default class ComponentWriter {
       return undefined;
     };
 
-    return this.consumer.bitMap.addComponent({
+    return this.bitMap.addComponent({
       componentId: this.component.id,
       files: filesForBitMap,
       mainFile: this.component.mainFile, // $FlowFixMe
@@ -220,6 +232,7 @@ export default class ComponentWriter {
   }
 
   _copyFilesIntoDistsWhenDistsOutsideComponentDir() {
+    if (!this.consumer) return; // not relevant when consumer is not available
     if (!this.consumer.shouldDistsBeInsideTheComponent() && this.component.dists.isEmpty()) {
       // since the dists are set to be outside the components dir, the source files must be saved there
       // otherwise, other components in dists won't be able to link to this component
@@ -245,14 +258,15 @@ export default class ComponentWriter {
    * never reach this function, see @write-components.writeToComponentsDir). Therefore, always add to bit.map.
    */
   _updateBitMapIfNeeded() {
-    const componentMapExistWithSameVersion = this.consumer.bitMap.isExistWithSameVersion(this.component.id);
+    if (this.isolated) return;
+    const componentMapExistWithSameVersion = this.bitMap.isExistWithSameVersion(this.component.id);
     const updateBitMap =
       !componentMapExistWithSameVersion || this.componentMap.originallySharedDir !== this.component.originallySharedDir;
     if (updateBitMap) {
       if (componentMapExistWithSameVersion) {
         // originallySharedDir has been changed. it affects also the relativePath of the files
         // so it's better to just remove the old record and add a new one
-        this.consumer.bitMap.removeComponent(this.component.id);
+        this.bitMap.removeComponent(this.component.id);
       }
       this.component.componentMap = this.addComponentToBitMap(this.componentMap.rootDir);
     }
@@ -263,8 +277,8 @@ export default class ComponentWriter {
     // so if the overrides or envs were changed, it should be written to the consumer-config
     const areEnvsChanged = async (): Promise<boolean> => {
       const context = { componentDir: this.componentMap.getRootDir() };
-      const compilerFromConsumer = await this.consumer.getEnv(COMPILER_ENV_TYPE, context);
-      const testerFromConsumer = await this.consumer.getEnv(TESTER_ENV_TYPE, context);
+      const compilerFromConsumer = this.consumer ? await this.consumer.getEnv(COMPILER_ENV_TYPE, context) : null;
+      const testerFromConsumer = this.consumer ? await this.consumer.getEnv(TESTER_ENV_TYPE, context) : null;
       const compilerFromComponent = this.component.compiler ? this.component.compiler.toModelObject() : null;
       const testerFromComponent = this.component.tester ? this.component.tester.toModelObject() : null;
       return (
@@ -278,7 +292,7 @@ export default class ComponentWriter {
         )
       );
     };
-    if (this.componentMap.origin === COMPONENT_ORIGINS.AUTHORED) {
+    if (this.componentMap.origin === COMPONENT_ORIGINS.AUTHORED && this.consumer) {
       this.consumer.config.overrides.updateOverridesIfChanged(this.component, await areEnvsChanged());
     }
   }
@@ -303,6 +317,7 @@ export default class ComponentWriter {
    * bit.map and add an updated one.
    */
   async _handlePreviouslyNestedCurrentlyImportedCase() {
+    if (!this.consumer) return;
     if (this.origin === COMPONENT_ORIGINS.IMPORTED && this.componentMap.origin === COMPONENT_ORIGINS.NESTED) {
       await this._cleanOldNestedComponent();
       this.componentMap = this.addComponentToBitMap(this.writeToPath);
@@ -330,6 +345,7 @@ export default class ComponentWriter {
   }
 
   async _cleanOldNestedComponent() {
+    if (!this.consumer) throw new Error('ComponentWriter._cleanOldNestedComponent expect to have a consumer');
     // $FlowFixMe this function gets called when it was previously NESTED, so the rootDir is set
     const oldLocation = path.join(this.consumer.getPath(), this.componentMap.rootDir);
     logger.debugAndAddBreadCrumb(
@@ -339,21 +355,35 @@ export default class ComponentWriter {
     );
     await fs.remove(oldLocation);
     await this._removeNodeModulesLinksFromDependents();
-    this.consumer.bitMap.removeComponent(this.component.id);
+    this.bitMap.removeComponent(this.component.id);
   }
 
   async _removeNodeModulesLinksFromDependents() {
+    if (!this.consumer) {
+      throw new Error('ComponentWriter._removeNodeModulesLinksFromDependents expect to have a consumer');
+    }
     const directDependentIds = await this.consumer.getAuthoredAndImportedDependentsIdsOf([this.component]);
     await Promise.all(
       directDependentIds.map((dependentId) => {
-        const dependentComponentMap = this.consumer.bitMap.getComponent(dependentId);
-        const relativeLinkPath = getNodeModulesPathOfComponent(this.consumer.config.bindingPrefix, this.component.id);
-        const nodeModulesLinkAbs = this.consumer.toAbsolutePath(
-          path.join(dependentComponentMap.getRootDir(), relativeLinkPath)
-        );
-        logger.debug(`deleting an obsolete link to node_modules at ${nodeModulesLinkAbs}`);
-        return fs.remove(nodeModulesLinkAbs);
+        const dependentComponentMap = this.consumer ? this.consumer.bitMap.getComponent(dependentId) : null;
+        const relativeLinkPath = this.consumer
+          ? getNodeModulesPathOfComponent(this.consumer.config.bindingPrefix, this.component.id)
+          : null;
+        const nodeModulesLinkAbs =
+          this.consumer && dependentComponentMap && relativeLinkPath
+            ? this.consumer.toAbsolutePath(path.join(dependentComponentMap.getRootDir(), relativeLinkPath))
+            : null;
+        if (nodeModulesLinkAbs) {
+          logger.debug(`deleting an obsolete link to node_modules at ${nodeModulesLinkAbs}`);
+        }
+        return nodeModulesLinkAbs ? fs.remove(nodeModulesLinkAbs) : Promise.resolve();
       })
     );
+  }
+
+  _getConfigDir() {
+    if (this.configDir) return this.configDir;
+    if (this.consumer) return this.consumer.dirStructure.ejectedEnvsDirStructure;
+    return new ConfigDir(DEFAULT_EJECTED_ENVS_DIR_PATH);
   }
 }