+++ src/links/index.js
@@ -1,15 +1,8 @@
 import NodeModuleLinker from './node-modules-linker';
-import {
-  linkComponents,
-  linkAllToNodeModules,
-  getLinksInDistToWrite,
-  getAllComponentsLinks,
-  reLinkDependents
-} from './linker';
+import { linkAllToNodeModules, getLinksInDistToWrite, getAllComponentsLinks, reLinkDependents } from './linker';
 import LinkFile from './link-file';
 
 export {
-  linkComponents,
   NodeModuleLinker,
   linkAllToNodeModules,
   getLinksInDistToWrite,