+++ src/utils/object/filter.js
@@ -4,9 +4,9 @@ import forEach from './foreach';
 /**
  * create a new array with all elements that pass the test implemented by the provided function.
  * @name filter
- * @param {object} obj object or array to iterate 
+ * @param {object} obj object or array to iterate
  * @param {function} cb callback function to invoke
- * @example 
+ * @example
  * ```js
  *  filter({ a: 1, b: 2, c: 3 }, (val, key) => val === 1) // => { a: 1 }
  * ```