+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- add option to isolate component into "capsule"
+- enable compilers to isolate components using capsule
+
 ## [14.1.4-angular.1] - 2019-06-10
 
 - support identify angular dependencies